+++ src/pallas/client.py
@@ -16,7 +16,6 @@ from __future__ import annotations
 
 import copy
 import time
-import warnings
 from typing import Iterable, Optional
 
 from pallas.caching import AthenaCache
@@ -282,8 +281,6 @@ class Athena:
         self,
         operation: str,
         parameters: PARAMETERS = None,
-        *,
-        ignore_cache: bool = False,
     ) -> QueryResults:
         """
         Execute a query and return results.
@@ -302,26 +299,14 @@ class Athena:
             All substitute parameters are quoted appropriately.
             See the :meth:`.quote` method for a supported parameter types.
         :type parameters: Union[None, Tuple[SQL_SCALAR, ...], Mapping[str, SQL_SCALAR]]
-        :param ignore_cache: deprecated, do not use.
         :return: query results
         """
-        if ignore_cache:
-            warnings.warn(
-                "The athena.execute(..., ignore_cache=True) parameter is deprecated."
-                " Use athena.using(cache_read=False).execute(...) instead.",
-                FutureWarning,
-            )
-            return self.using(cache_read=False).execute(operation, parameters)
-        return self.submit(
-            operation, parameters, ignore_cache=ignore_cache
-        ).get_results()
+        return self.submit(operation, parameters).get_results()
 
     def submit(
         self,
         operation: str,
         parameters: PARAMETERS = None,
-        *,
-        ignore_cache: bool = False,
     ) -> Query:
         """
         Submit a query and return.
@@ -340,16 +325,8 @@ class Athena:
             All substitute parameters are quoted appropriately.
             See the :meth:`.quote` method for a supported parameter types.
         :type parameters: Union[None, Tuple[SQL_SCALAR, ...], Mapping[str, SQL_SCALAR]]
-        :param ignore_cache: deprecated, do not use.
         :return: a query instance
         """
-        if ignore_cache:
-            warnings.warn(
-                "The athena.submit(..., ignore_cache=True) parameter is deprecated."
-                " Use athena.using(cache_read=False).submit(...) instead.",
-                FutureWarning,
-            )
-            return self.using(cache_read=False).submit(operation, parameters)
         sql = self._get_sql(operation, parameters)
         should_cache = is_select(sql)
         if should_cache: